move rgbToHex to utils
